+++ src/core/queue/merge/spec.js
@@ -57,4 +57,24 @@ describe('core/queue/merge', () => {
 		expect(queue.head).toBeUndefined();
 		expect(queue.length).toBe(0);
 	});
+
+	it('cloning a queue', () => {
+		const
+			queue = new MergeQueue();
+
+		queue.push(3);
+
+		const
+			clonedQueue = queue.clone();
+
+		expect(queue !== clonedQueue).toBe(true);
+
+		expect(queue.head).toBe(0);
+		expect(queue.length).toBe(1);
+		expect(queue.pop()).toBe(3);
+
+		expect(clonedQueue.head).toBe(0);
+		expect(clonedQueue.length).toBe(1);
+		expect(clonedQueue.pop()).toBe(3);
+	});
 });