+++ src/core/queue/order/spec.js
@@ -55,4 +55,26 @@ describe('core/queue/order', () => {
 		expect(queue.head).toBeUndefined();
 		expect(queue.length).toBe(0);
 	});
+
+	it('cloning a queue', () => {
+		const
+			queue = new OrderQueue((a, b) => a - b);
+
+		queue.push(3);
+
+		const
+			clonedQueue = queue.clone();
+
+		expect(queue !== clonedQueue).toBe(true);
+
+		expect(queue.head).toBe(0);
+		expect(queue.length).toBe(1);
+		expect(queue.pop()).toBe(3);
+
+		expect(clonedQueue.head).toBe(0);
+		expect(clonedQueue.length).toBe(1);
+		expect(clonedQueue.pop()).toBe(3);
+
+		expect(queue.comparator === clonedQueue.comparator).toBe(true);
+	});
 });