+++ src/core/queue/worker/merge/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.? (2022-0?-??)
+
+#### :rocket: New Feature
+
+* Added new `clone` method
+
 ## v3.20.0 (2020-07-05)
 
 #### :rocket: New Feature